// src/types/eventDetail.types.ts
export interface EventDetail {
  user: string;
  streamType: string;
  title: string;
  overrides: any;
  category: "past" | "present" | "future";
  vector: string;
  duration: number;
  description: string;
  startDate: string;
  endDate: string;
  imageUrl: string;
  validFrom: string;
  validTo: string;
  childEvents: any;
  med: boolean;
  channel: any;
  tags: any;
  highlightOffset: number;
  sport: string;
  videoSrc: string;
  eventItemId: string;
  cesimId: any;
  streamId: any;
  contest: string;
  client: string;
  urn: string;
  hls: string;
  meta: any;
  system: string;
  integration: any;
  geoblock: boolean;
  firstLiveSegmentStartTime: string;
  segmentDuration: number;
  dvrWindow: number;
  id: string;
  createdDate: string;
  createdBy: string;
  changedDate: string;
  changedBy: string;
  analyticsMetadata: {
    media_urn: string;
    media_segment_length: number;
    media_episode_length: number;
    media_episode: string;
    media_show: string;
    media_special: string;
    media_segment: string;
    media_full_length: string;
    media_is_geoblocked: boolean;
    media_url: string;
    media_channel_name: string;
    media_channel_id: string;
    media_channel_code: string;
    media_livestream_encoder: any;
    media_content_group: string;
    media_episode_id: string;
    media_publication_date: string;
    media_publication_time: string;
    media_is_tvsvizzera: string;
    media_special_format: any;
    media_quality: any;
    media_since_publication_d: number;
    media_since_publication_h: number;
    media_thumbnail: string;
    media_type: string;
    media_is_livestream: string;
    media_is_web_only: any;
    media_show_id: string;
    media_publication_datetime: string;
    media_camera_angle_name: any;
    media_language: string;
    media_enterprise_units: string;
    media_joker1: string;
    media_joker2: string;
    media_joker3: string;
  }
}